% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apa_print_mlsdt.R
\name{apa_print_mlsdt}
\alias{apa_print_mlsdt}
\title{Fit a multilevel signal detection theory model}
\usage{
apa_print_mlsdt(fit, tests, coef, coef_pretty = NULL, direction = NULL, CI = T)
}
\arguments{
\item{fit}{fitted glmer() model}

\item{tests}{object returned by compute_tests(), including the parameter in question}

\item{coef}{internal name of the coefficient (name in the glmer fit)}

\item{coef_pretty}{Subscript of the coefficient for the formatted output}
}
\value{
list
}
\description{
Fit a multilevel signal detection theory model
}
