% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_lrts.R
\name{compute_LRTs}
\alias{compute_LRTs}
\title{Compute Likelihood Ratio Tests for a fitted SDT model}
\usage{
compute_LRTs(
  fit_obj = NULL,
  formula_mu,
  formula_lambda,
  dv,
  data,
  mm = NULL,
  type = 3,
  test_intercepts = F,
  test_ran_ef = F,
  correlate_sdt_params = T
)
}
\arguments{
\item{fit_obj}{An lme4 or glmmTMB fit object containing the full fit and
coefficients that should be tested}

\item{formula_mu}{the corresponding formula for sensitivity}

\item{formula_lambda}{the corresponding formula for response bias}

\item{dv}{the name of the dependent variable in the data}

\item{data}{a data frame}

\item{mm}{model matrices (optional)}

\item{type}{type of tests (only relevant for likelihood ratio tests and
parametric bootstrapping)}

\item{test_intercepts}{boolean indicating if intercepts for sensitivity and
response bias should be tested}

\item{test_ran_ef}{boolean indicating whether random (T) or fixed effects
should be tested (F)}

\item{correlate_sdt_params}{boolean indicating if correlations between SDT
parameters should be estimated}
}
\description{
Compute Likelihood Ratio Tests for a fitted SDT model
}
