% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_glmer_formula.R
\name{construct_glmer_formula}
\alias{construct_glmer_formula}
\title{Convert two given formulas for the two SDT parameters to a glmer() formula}
\usage{
construct_glmer_formula(
  formula_mu,
  formula_lambda,
  dv,
  correlate_sdt_params = TRUE,
  mm = NULL,
  to_remove = NULL,
  remove_correlations = FALSE,
  test_params_mu = "all",
  test_params_lambda = "all"
)
}
\arguments{
\item{formula_mu}{formula for discriminability}

\item{formula_lambda}{formula for response bias}

\item{dv}{name of dependent variable}

\item{correlate_sdt_params}{model correlations between mu and lambda random effects?}

\item{mm}{model matrices (corresponding to the formulas) -> necessary for formulas with uncorrelated
random effects and for indices to be removed}

\item{to_remove}{vector of indices that should be removed from the formula}

\item{remove_correlations}{boolean indicating whether correlations should be removed from the formula}

\item{test_params_mu}{which coefficients for sensitivity should be tested}

\item{test_params_lambda}{which coefficients for response bias should be tested}
}
\value{
lme4 formula
}
\description{
Convert two given formulas for the two SDT parameters to a glmer() formula
}
