% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mlsdt.R
\name{construct_modeldata}
\alias{construct_modeldata}
\title{Construct Model Matrices From the Given Data to give to glmer()}
\usage{
construct_modeldata(
  formula_mu,
  formula_lambda,
  dv,
  data,
  trial_type_var = "trial_type"
)
}
\arguments{
\item{formula_mu}{formula for sensitivity (mu)}

\item{formula_lambda}{formula for response bias (lambda)}

\item{dv}{name of dependent variable}

\item{data}{dataset used to construct the model data}

\item{trial_type_var}{name of variable coding the type of trial (signal vs. noise)}
}
\value{
list of 4 model matrices (fixed and random for mu and lambda)
}
\description{
Construct Model Matrices From the Given Data to give to glmer()
}
\examples{
construct_modeldata(
  formula_mu = ~ x1 + (1 | ID)
  formula_lambda = ~ x1 + (1 | ID)
  data = data
)
}
