% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_modelmatrices.R
\name{construct_modelmatrices}
\alias{construct_modelmatrices}
\title{Construct Model Matrices From the Given Data to give to glmer()}
\usage{
construct_modelmatrices(
  formula_mu,
  formula_lambda,
  data,
  trial_type_var = "trial_type",
  distribution = "gaussian"
)
}
\arguments{
\item{formula_mu}{formula for sensitivity (mu)}

\item{formula_lambda}{formula for response bias (lambda)}

\item{data}{dataset used to construct the model data}

\item{trial_type_var}{name of variable coding the type of trial (signal vs. noise)}
}
\value{
list of model matrices (fixed and random for mu and lambda)
}
\description{
Construct Model Matrices From the Given Data to give to glmer()
}
