% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emmeans.mesdt_fit.R
\name{emmeans.mesdt_fit}
\alias{emmeans.mesdt_fit}
\title{Estimated Marginal Means for mesdt_fit Objects}
\arguments{
\item{object}{A \code{mesdt_fit} object.}

\item{specs}{A \code{character} vector specifying the names of the predictors
over which EMMs are desired. specs may also be a \code{formula} or a \code{list}
(optionally named) of valid specs. Use of formulas is described in the
Overview section below. Note: We recommend against using two-sided formulas;
see the note below for contr.}

\item{dpar}{Character string specifying the distributional parameter
(e.g., \code{"mu"}, \code{"sigma"}) for which the EMMs are computed.
If \code{NULL} (the default), the location/mean parameter is used.}

\item{...}{Additional arguments passed to \code{emmeans}. See ?emmeans for
details on additional arguments to emmeans.}
}
\value{
An object of class \code{emmGrid}, containing the estimated marginal means
for the specified SDT parameter.
}
\description{
Compute estimated marginal means (EMMs) for objects of class \code{mymodel}
for discriminability or response bias.
}
\examples{
\dontrun{
mod_mixed <- fit_mesdt(
  discriminability ~ committee * emp_gender + (1 | id),
  response_bias ~ committee * emp_gender + (committee | id),
  data = debi3subset,
  trial_type = "status",
  dv = "assessment"
)
em_response_bias <- emmeans(mod_mixed, specs = ~ committee, dpar = "response bias")
em_discriminability <- emmeans(mod_mixed, specs = ~ committee, dpar = "discriminability")
}

}
