% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mlsdt.R
\name{fit_glmm}
\alias{fit_glmm}
\title{Fit the GLMM}
\usage{
fit_glmm(formula, data, backend = "lme4")
}
\arguments{
\item{formula}{The glmer model formula}

\item{data}{data the model should be fitted to}

\item{backend}{whether to use R (lme4) or Julia (MixedModels)}
}
\value{
lme4 fit
}
\description{
Fit the GLMM
}
