% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mlsdt.R
\name{fit_mlsdt}
\alias{fit_mlsdt}
\title{Fit a multilevel signal detection theory model}
\usage{
fit_mlsdt(
  formula_mu,
  formula_lambda,
  dv,
  trial_type_var = "trial_type",
  data,
  correlate_sdt_params = T,
  max_iter = 1e+06,
  fast = T
)
}
\arguments{
\item{formula_mu}{Formula specifying fixed and random effects on sensitivity}

\item{formula_lambda}{Formula specifying fixed and random effects on response bias}

\item{dv}{name of the (binary) dependent variable}

\item{trial_type_var}{name of the variable coding signal vs. noise trials}

\item{data}{dataset}

\item{backend}{package / library used to fit the model
-> only supports R lme4 (backend = "lme4" at the moment but support for Julia
MixedModels and glmmTMB is planned)}
}
\value{
TODO
}
\description{
Fit a multilevel signal detection theory model
}
