% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mlsdt.R
\name{make_glmer_formula}
\alias{make_glmer_formula}
\title{Convert two given formulas for the two SDT parameters to an lme4 formula}
\usage{
make_glmer_formula(
  form_mu,
  form_lambda,
  dv,
  trial_type_var = "trial_type",
  random,
  between = NULL,
  within = NULL
)
}
\arguments{
\item{form_mu}{formula for sensitivity}

\item{form_lambda}{formula for response bias}

\item{dv}{name of dependent variable}

\item{trial_type_var}{name of variable coding the type of trial (signal vs. noise)}

\item{random}{name of the variable of the random factor}

\item{between}{predictors varying between levels of the random factor}

\item{within}{predictors varying within levels of the random factor}
}
\value{
lme4 formula
}
\description{
Convert two given formulas for the two SDT parameters to an lme4 formula
}
\examples{
make_glmer_formula(
  form_mu = mu ~ x_test,
  form_lambda = lambda ~ x_test,
  dv = "Y",
  trial_type_var = "trial_type",
  random = "ID",
  within = c("x_test")
 )
}
