% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesdt_fit_class.R
\name{print.mesdt_fit}
\alias{print.mesdt_fit}
\title{Print Method for Summary of mesdt_fit Objects}
\usage{
\method{print}{summary.mesdt_fit}(x, ...)
}
\arguments{
\item{x}{An object of class \code{summary.mesdt_fit}, typically returned by
\code{summary.mesdt_fit()}.}

\item{...}{Additional arguments passed to or from other methods. Currently unused.}
}
\value{
Invisibly returns \code{NULL}. The function prints results to the console.
}
\description{
Prints a structured summary of a \code{summary.mesdt_fit} object.
This method allows formatting options such as the number of significant
digits and whether to show significance stars.
Alias for print.summary.mesdt_fit
}
