% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{set_backend}
\alias{set_backend}
\title{Set backend for mesdt.}
\usage{
set_backend(backend_name)
}
\arguments{
\item{backend_name}{\code{character} indicating the backend that should be used
to estimate the models (\code{lme4}, the default, which is set when loading the
package, or \code{glmmTMB}).}
}
\description{
Allows the user to change the package that is used to estimate the
models. When loading the package, the backend is automatically set to \code{lme4}
(which is a dependency for this package). If \code{glmmTMB} is installed, the
backend can be changed to that package through this function.
}
\examples{
set_backend("lme4")
\dontrun{
set_backend("glmmTMB")
}
}
