% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesdt_fit_class.R
\name{summary.mesdt_fit}
\alias{summary.mesdt_fit}
\alias{print.summary.mesdt_fit}
\title{Summary Method for mesdt_fit Objects}
\usage{
\method{summary}{mesdt_fit}(object, ...)

\method{print}{summary.mesdt_fit}(x, digits = max(3, getOption("digits") - 3), signif.stars = FALSE, ...)
}
\arguments{
\item{object}{An object of class \code{mesdt_fit}, typically returned by a
model fitting function such as \code{mesdt()}.}

\item{...}{Additional arguments passed to or from other methods. Currently unused.}

\item{x}{An object of class \code{summary.mesdt_fit}.}

\item{digits}{Integer specifying the number of significant digits to print.
Defaults to \code{max(3, getOption("digits") - 3)}.}

\item{signif.stars}{Logical; if \code{TRUE}, significance stars are displayed
next to p-values. Defaults to \code{FALSE}.}
}
\value{
A summary object for \code{mesdt_fit} similar to the summary provided
by lme4.

Invisibly returns \code{NULL}. The function prints results to the console.
}
\description{
Provides a summary of a \code{mesdt_fit} object. Currently, this function
is a placeholder and can be extended to include likelihood ratio tests,
bootstrap results, or other relevant summary statistics.

Estimates (mixed-effects) signal detection theory (SDT) models with maximum
likelihood estimation by leveraging the equivalence between certain SDT and
certain generalized linear models (GLMM; De Carlo, 1998). The GLMMs are
estimated using either the \code{lme4} (the default) or the \code{glmmTMB} package
(which can be significantly faster) as a backend, which can be changed with
\link{set_backend}.

The parameter estimates can be interpreted on the latent evidence
scale: For instance, when using sum contrasts, the fixed (or population-level)
intercepts for discriminability (d') and response bias (c) indicate mean overall
levels of discriminability and response bias. The higher d', the larger the
difference between the means of the signal and noise distributions and the
better signal and noise can be discriminated. A response bias of c < 0 (c > 0)
indicates a liberal (conservative) response tendency, that is, a tendency
to give a "signal" ("noise") response. The fixed effects of predictors
(again, assuming sum contrasts for all predictors) describe the population
effect of an predictor on discriminability or response bias. (i.e., for a
categorical variable with two levels, the difference between those levels).
The random effects indicate the variability between the levels of the
random-effects grouping factor (e.g., between participants).
For a more detailed explanation and tutorial see Jakob et al. (2025).
}
\details{
Prints a structured summary of a \code{summary.mesdt_fit} object.
This method allows formatting options such as the number of significant
digits and whether to show significance stars.
}
