% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mlsdt.R
\name{transform_to_sdt}
\alias{transform_to_sdt}
\title{Title}
\usage{
transform_to_sdt(fit_obj, formula_lambda, formula_mu, data, level = 0.95)
}
\arguments{
\item{data}{}
}
\value{
TODO: test this...
}
\description{
Title
}
