% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_tests.R
\name{compute_tests}
\alias{compute_tests}
\title{Compute Likelihood Ratio Tests or tests based on parametric bootstrapping
for a fitted SDT model}
\usage{
compute_tests(
  mesdt_fit,
  cl = NULL,
  tests = "lrt",
  nsim = 1000,
  type = 3,
  test_intercepts = F,
  test_ran_ef = F,
  tests_discriminability = "all",
  tests_response_bias = "all",
  control = NULL,
  seed = NULL
)
}
\arguments{
\item{tests}{type of tests that should be computed ("LRT" -> likelihood ratio tests,
"bootstrap" = parametric bootstrap)}

\item{nsim}{number of simulated datasets for bootstrapping}

\item{type}{type of tests (2 or 3, only relevant for likelihood ratio tests and
parametric bootstrapping)}

\item{test_intercepts}{boolean indicating if intercepts for discriminability and
response bias should be tested}

\item{test_ran_ef}{boolean indicating whether random (T) or fixed effects
should be tested (F)}

\item{fit_obj}{An mesdt fit object containing the full fit and
coefficients that should be tested}

\item{test_discriminability}{which coefficients on discriminability should be tested?
(default is "all")}

\item{test_response_bias}{which coefficients on response bias should be tested?
(default is "all")}
}
\description{
Compute Likelihood Ratio Tests or tests based on parametric bootstrapping
for a fitted SDT model
}
