% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mlsdt.R
\name{construct_glmer_formula}
\alias{construct_glmer_formula}
\title{Convert two given formulas for the two SDT parameters to a glmer() formula}
\usage{
construct_glmer_formula(
  formula_mu,
  formula_lambda,
  dv,
  param_idx = NULL,
  remove_from_mu = NULL
)
}
\arguments{
\item{formula_mu}{formula for sensitivity}

\item{formula_lambda}{formula for response bias}

\item{dv}{name of dependent variable}

\item{param_idx}{optional index of parameter to be removed to construct a reduced formula}

\item{remove_from_mu}{optional argument to indicate whether the to-be-removed parameter
should be removed from mu or lambda modeldata}

\item{trial_type_var}{name of variable coding the type of trial (signal vs. noise)}
}
\value{
lme4 formula
}
\description{
Convert two given formulas for the two SDT parameters to a glmer() formula
}
\examples{
construct_glmer_formula(
  formula_mu = ~ x1 + (1 | ID)
  formula_lambda = ~ x2 + (1 | ID)
  dv = "y",
  trial_type_var = "trial_type"
)
}
