% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_glmer_formula.R
\name{construct_glmer_formula}
\alias{construct_glmer_formula}
\title{Convert two given formulas for the two SDT parameters to a glmer() formula}
\usage{
construct_glmer_formula(
  formula_mu,
  formula_lambda,
  dv,
  correlate_sdt_params = T,
  mm = NULL,
  param_idc = NULL,
  remove_from_mu = NULL
)
}
\arguments{
\item{formula_mu}{formula for sensitivity}

\item{formula_lambda}{formula for response bias}

\item{dv}{name of dependent variable}

\item{correlate_sdt_params}{model correlations between mu and lambda random effects?}

\item{mm}{model matrices (corresponding to the formulas) -> only necessary for formulas with uncorrelated random effects}

\item{param_idc}{optional vector of parameters indices to be removed to construct a reduced formula}

\item{remove_from_mu}{optional argument to indicate whether the to-be-removed parameter
should be removed from mu or lambda model matrix}
}
\value{
lme4 formula
}
\description{
Convert two given formulas for the two SDT parameters to a glmer() formula
}
\examples{
construct_glmer_formula(
  formula_mu = ~ x1 + (1 | ID)
  formula_lambda = ~ x2 + (1 | ID)
  dv = "y",
  trial_type_var = "trial_type"
)
}
