% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mesdt.R
\name{fit_mesdt}
\alias{fit_mesdt}
\title{Fit a (mixed-effects) signal detection theory (SDT) model}
\usage{
fit_mesdt(
  discriminability,
  response_bias,
  dv,
  trial_type,
  data,
  distribution = c("gaussian", "logistic", "gumbel-min", "gumbel-max"),
  correlate_sdt_params = FALSE,
  aggregate = FALSE,
  control = NULL
)
}
\arguments{
\item{response_bias}{\code{formula} specifying fixed and random effects on response bias
with the common syntax for mixed-effects models from lme4 (see Details).}

\item{dv}{\code{character} string specifying the name of the (binary) dependent
variable.}

\item{trial_type}{\code{character} specifying name of the variable coding
whether a trial in the given df is a signal or a noise trial.}

\item{data}{\code{data.frame} containing the variables used in the formulas for
discriminability and response bias.}

\item{distribution}{\code{character} specifying the parametric distribution
of signal and noise evidence ("gaussian", which is the default, "logistic",
or "gumbel-min").}

\item{correlate_sdt_params}{\code{boolean} indicating whether correlations between
SDT parameters should be modeled (default is FALSE, see Details).}

\item{aggregate}{(experimental) \code{boolean} indicating whether the given long
data frame should be aggregated. Can significantly speed up estimation,
especially for lme4, but has the disadvantage that additional methods based
on the fitted model (e.g., bootstrapping) might not work as expected.}

\item{control}{list containing optional control arguments that are included
in the \code{glmmTMB()} or \code{glmer()} call (see Details).}

\item{formula}{\code{formula} specifying fixed and random effects on
discriminability with the common syntax for mixed-effects models from
\code{lme4} (see Details).}
}
\value{
An object of class \code{mesdt_fit}, containing the fitted model
(\verb{$fit_obj}), information about the specified strucure of the model
(\verb{$user_input}) and internal information used for post-processing.
(\verb{$internal}).
}
\description{
Estimates (mixed-effects) signal detection theory (SDT) models with maximum
likelihood estimation by leveraging the equivalence between certain SDT and
certain generalized linear models (GLMM; De Carlo, 1998). The GLMMs are
estimated using either the \code{lme4} (the default) or the \code{glmmTMB} package
(which can be significantly faster) as a backend, which can be changed with
\link{set_backend}.

The resulting parameter estimates can be interpreted on the latent evidence
scale that is assumed in SDT: For instance, when using sum contrasts, the fixed (or population-level)
intercepts for discriminability (d') and response bias (c) indicate mean overall
levels of discriminability and response bias. The higher d', the larger the
difference between the means of the signal and noise distributions and the
better signal and noise can be discriminated. A response bias of c < 0 (c > 0)
indicates a liberal (conservative) response tendency, that is, a tendency
to give a "signal" ("noise") response. The fixed effects of predictors
(again, assuming sum contrasts for all predictors) describe the population
effect of an predictor on discriminability or response bias. (i.e., for a
categorical variable with two levels, the difference between those levels).
The random effects indicate the variability between the levels of the
random-effects grouping factor (e.g., between participants).
For a more detailed explanation and tutorial see Jakob et al. (2025).

The default SDT parametrization is the common equal-variance Gaussian model,
but \code{fit_mesdt()} and this package support other distributions (i.e., the
logistic, gumbel-min and gumbel-max distribution) as well, which can be
more appropriate in certain contexts.

\code{summary()} and \code{print()} methods are provided, showing parameter estimates,
standard errors and results of Wald tests (as returned by \code{lme4} and
\code{glmmTMB}); however, it is usually recommended to rely on likelihood ratio
tests or parametric bootstrapping tests (in case of few levels of the
random-effects grouping factor) for statistical inference with GLMM, which
are provided in the \link{compute_tests} function.

In addition, the returned \code{mesdt_fit} object contains the fitted GLMM,
(\verb{$fit_obj}, a \code{glmerfit}), allowing users to do additional processing that is not (yet)
implemented within this package, such as post-processing of random effect
conditional modes.
}
\details{
\itemize{
\item Formulas for discriminability and response bias should be one-sided (e.g.,
\code{discriminability =~ 1 + (1 | id)}).
\item Suppressing correlations between random effects is possible by dividing
the random-effects formula and the grouping factor with two, instead of one
vertical bars (\code{||}). Note that unlike lme4 and glmmTMB, mesdt also supports
this notation for factors.
\item The current implementation supports completely correlated and
completely uncorrelated random-effects structures, as well as correlations
only within random effects for the same SDT parameter (i.e., random effects
for discriminability are correlated with each other but not with random
effects for response bias and vice versa). Formulas containing both
correlated and uncorrelated terms are not yet supported and will be estimated
as completely uncorrelated models.
\item Formulas can also specify single-level models without random effects. In
such a case, the \code{glm()} function from \code{stats} is used to estimate the model.
\item Per default, correlations between the random effects for discriminability
and response bias are modeled with a joint covariance matrix (i.e., random
intercepts and slopes for discriminability and response bias are correlated
with each other). Setting \code{correlate_sdt_params = F} splits this up into
two separate covariance matrices for discriminability and response bias
(which can help to achieve convergence).
\item The \code{control} argument allows to pass all control arguments taken by
\code{lme4} or \code{glmmTMB} (depending on the backend used; see the relevant
documentation for details).
}
}
\examples{
\dontrun{
# Mixed-effects SDT model
# by-participant random intercepts for msensitivity and response bias
# by-participant random slope for the effect of the committee decision on
# response bias
# correlations between all random effects
mod_mixed <- fit_mesdt(
  discriminability ~ committee * emp_gender + (1 | id),
   response_bias ~ committee * emp_gender + (committee | id),
   data = debi3_sub,
   trial_type = "status",
   dv = "assessment"
)
summary(mod_mixed)

# Fixed-effects model (not recommended for this type of multi-level data structure!):
mod_fixed_only <- fit_mesdt(
  discriminability ~ committee * emp_gender,
  response_bias ~ committee * emp_gender,
   data = debi3_sub,
   trial_type = "status",
   dv = "assessment"
)
summary(mod_fixed_only)
}
}
