% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mesdt.R
\name{fit_mesdt}
\alias{fit_mesdt}
\title{Fit a multilevel signal detection theory model}
\usage{
fit_mesdt(
  discriminability,
  bias,
  dv,
  trial_type_var = "trial_type",
  data,
  distribution = "gaussian",
  correlate_sdt_params = T,
  control = NULL
)
}
\arguments{
\item{discriminability}{Formula specifying fixed and random effects on sensitivity}

\item{bias}{Formula specifying fixed and random effects on response bias}

\item{dv}{name of the (binary) dependent variable}

\item{trial_type_var}{name of the variable coding signal vs. noise trials}

\item{data}{dataset}

\item{correlate_sdt_params}{boolean indicating whether correlations between
SDT parameters should be modeled}
}
\value{
TODO
}
\description{
Fit a multilevel signal detection theory model
}
