% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mlsdt.R
\name{fit_mlsdt}
\alias{fit_mlsdt}
\title{Fit a multilevel signal detection theory model}
\usage{
fit_mlsdt(
  formula_mu,
  formula_lambda,
  dv,
  trial_type_var = "trial_type",
  data,
  correlate_sdt_params = T,
  tests = "Wald",
  control = NULL
)
}
\arguments{
\item{formula_mu}{Formula specifying fixed and random effects on sensitivity}

\item{formula_lambda}{Formula specifying fixed and random effects on response bias}

\item{dv}{name of the (binary) dependent variable}

\item{trial_type_var}{name of the variable coding signal vs. noise trials}

\item{data}{dataset}

\item{correlate_sdt_params}{boolean indicating whether correlations between
SDT parameters should be modeled}

\item{tests}{type of statistical tests that should be used ("Wald": Wald tests,
"LRT": likelihood ratio tests, "boot": parametric bootstrapping)}
}
\value{
TODO
}
\description{
Fit a multilevel signal detection theory model
}
