% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{get_backend}
\alias{get_backend}
\title{Get the backend for mesdt.}
\usage{
get_backend(backend_name)
}
\value{
the current backend used by mesdt.
}
\description{
Allows the user to check which package is currently set as the backend
to the estimate the models.
}
\examples{
set_backend()
}
